Implement base64-encoded data: URLs
Either percent- or base64-decode the URL text, as appropriate.  We now
always malloc a (decoded) copy of the url argument, rather than somewhat
illegitimately holding on to the pointer provided, as previously.

Add percent- and base64-decoding functions.  When we are happy with their
signatures (perhaps they should have a way to malloc the output buffer
themselves?), we may wish to move their declarations to a public header.

Implements and fixes #422.

[NEWS]
* Data URLs ("data:,text") now follow the standard RFC 2397 format and may
  be base64-encoded (when written as "data:;base64,text") or may include
  percent-encoded characters.  HTSlib's previous over-simplified "data:text"
  format is no longer supported -- you will need to add an initial comma.